gas/makeTitle/プレスリリース/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
pytest==8.3.2
